Flesh out tick on world to work for dead cells too
- To handle shortcoming where we only iterate over cells that are in
the living cell we needed to incorporate some way of iterating over
cells that are dead as well. One thought was to rely on getting a
grid location (all possible locations) and then a sort location
(which will cover living cells), and then we should be able to
iterate over all.